refactor : answer delete

+++ src/main/java/nextstep/qna/domain/Answer.java
@@ -1,5 +1,6 @@
 package nextstep.qna.domain;
 
+import nextstep.qna.CannotDeleteException;
 import nextstep.qna.NotFoundException;
 import nextstep.qna.UnAuthorizedException;
 import nextstep.users.domain.NsUser;
@@ -52,14 +53,24 @@ public class Answer {
         return this.writer.equals(writer);
     }
 
-    DeleteHistory delete() {
-        //
-        this.deleted = true;
-        return new DeleteHistory(ContentType.ANSWER, id, writer, LocalDateTime.now());
+    DeleteHistory delete(NsUser user) {
+        try {
+            validateOwner(user);
+            this.deleted = true;
+            return new DeleteHistory(ContentType.ANSWER, id, writer, LocalDateTime.now());
+        } catch (CannotDeleteException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     public void toQuestion(Question question) {
         this.question = question;
+    }
+
+    private void validateOwner(NsUser loginUser) throws CannotDeleteException {
+        if (!this.isOwner(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
     }
 
     public Long getId() {
@@ -70,9 +81,6 @@ public class Answer {
         return writer;
     }
 
-    public String getContents() {
-        return contents;
-    }
 
     @Override
     public String toString() {
